feat: create sebulba config and add computation config
