+++ mava/configs/arch/sebulba.yaml
@@ -1,0 +1,24 @@
+# --- Sebulba config ---
+arch_name: "sebulba"
+num_envs: 16  # number of envs per thread
+
+# --- Evaluation ---
+evaluation_greedy: False # Evaluate the policy greedily. If True the policy will select
+  # an action which corresponds to the greatest logit. If false, the policy will sample
+  # from the logits.
+num_eval_episodes: 32 # Number of episodes to evaluate per evaluation.
+num_evaluation: 200 # Number of evenly spaced evaluations to perform during training.
+absolute_metric: True # Whether the absolute metric should be computed. For more details
+  # on the absolute metric please see: https://arxiv.org/abs/2209.10485
+
+# --- Sebulba devices config ---
+n_threads_per_executor: 1  # num of different threads/env batches per actor
+executor_device_ids: [0] # ids of actor devices
+learner_device_ids: [0] # ids of learner devices
+
+# --- Sebulba rollout and env config ---
+concurrency: False # whether actor and learner should run concurrently
+async_envs: True # "whether to use async vector or sync vector envs"
+
+# --- To be defined during training ---
+log_frequency: ~
