+++ mava/systems/anakin/ff_mappo.py
@@ -11,7 +11,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
 import copy
 import time
 from typing import Any, Dict, Tuple
@@ -31,35 +30,37 @@ from rich.pretty import pprint
 
 from mava.evaluator import anakin_evaluator_setup as evaluator_setup
 from mava.logger import Logger
-from mava.networks import RecurrentActor as Actor
-from mava.networks import ScannedRNN, get_networks
+from mava.networks import FeedForwardActor as Actor
+from mava.networks import get_networks
 from mava.types import (
+    ActorApply,
+    CriticApply,
     ExperimentOutput,
-    HiddenStates,
     LearnerFn,
+    LearnerState,
+    ObservationGlobalState,
     OptStates,
     Params,
     PPOTransition,
-    RecActorApply,
-    RecCriticApply,
-    RNNLearnerState,
 )
 from mava.utils.checkpointing import Checkpointer
+from mava.utils.jax import merge_leading_dims
 from mava.utils.make_env import make
 
 
 def get_learner_fn(
     env: Environment,
-    apply_fns: Tuple[RecActorApply, RecCriticApply],
+    apply_fns: Tuple[ActorApply, CriticApply],
     update_fns: Tuple[optax.TransformUpdateFn, optax.TransformUpdateFn],
     config: DictConfig,
-) -> LearnerFn[RNNLearnerState]:
+) -> LearnerFn[LearnerState]:
     """Get the learner function."""
 
+    # Unpack apply and update functions.
     actor_apply_fn, critic_apply_fn = apply_fns
     actor_update_fn, critic_update_fn = update_fns
 
-    def _update_step(learner_state: RNNLearnerState, _: Any) -> Tuple[RNNLearnerState, Tuple]:
+    def _update_step(learner_state: LearnerState, _: Any) -> Tuple[LearnerState, Tuple]:
         """A single update of the network.
 
         This function steps the environment and records the trajectory batch for
@@ -74,57 +75,24 @@ def get_learner_fn(
                 - rng (PRNGKey): The random number generator state.
                 - env_state (State): The environment state.
                 - last_timestep (TimeStep): The last timestep in the current trajectory.
-                - dones (bool): Whether the last timestep was a terminal state.
-                - hstates (HiddenStates): The current hidden states of the RNN.
             _ (Any): The current metrics info.
         """
 
-        def _env_step(
-            learner_state: RNNLearnerState, _: Any
-        ) -> Tuple[RNNLearnerState, PPOTransition]:
+        def _env_step(learner_state: LearnerState, _: Any) -> Tuple[LearnerState, PPOTransition]:
             """Step the environment."""
-            (
-                params,
-                opt_states,
-                rng,
-                env_state,
-                last_timestep,
-                last_done,
-                hstates,
-            ) = learner_state
+            params, opt_states, rng, env_state, last_timestep = learner_state
 
+            # SELECT ACTION
             rng, policy_rng = jax.random.split(rng)
-
-            # Add a batch dimension to the observation.
-            batched_observation = jax.tree_util.tree_map(
-                lambda x: x[jnp.newaxis, :], last_timestep.observation
-            )
-            ac_in = (
-                batched_observation,
-                last_done[:, 0][jnp.newaxis, :],
-            )
-
-            # Run the network.
-            policy_hidden_state, actor_policy = actor_apply_fn(
-                params.actor_params, hstates.policy_hidden_state, ac_in
-            )
-            critic_hidden_state, value = critic_apply_fn(
-                params.critic_params, hstates.critic_hidden_state, ac_in
-            )
-
-            # Sample action from the policy and squeeze out the batch dimension.
+            actor_policy = actor_apply_fn(params.actor_params, last_timestep.observation)
+            value = critic_apply_fn(params.critic_params, last_timestep.observation)
             action = actor_policy.sample(seed=policy_rng)
             log_prob = actor_policy.log_prob(action)
-            value, action, log_prob = (
-                value.squeeze(0),
-                action.squeeze(0),
-                log_prob.squeeze(0),
-            )
 
-            # Step the environment.
+            # STEP ENVIRONMENT
             env_state, timestep = jax.vmap(env.step, in_axes=(0, 0))(env_state, action)
 
-            # log episode return and length
+            # LOG EPISODE METRICS
             done = jax.tree_util.tree_map(
                 lambda x: jnp.repeat(x, config.system.num_agents).reshape(config.arch.num_envs, -1),
                 timestep.last(),
@@ -137,14 +105,8 @@ def get_learner_fn(
             transition = PPOTransition(
                 done, action, value, timestep.reward, log_prob, last_timestep.observation, info
             )
-            hstates = HiddenStates(policy_hidden_state, critic_hidden_state)
-            learner_state = RNNLearnerState(
-                params, opt_states, rng, env_state, timestep, done, hstates
-            )
+            learner_state = LearnerState(params, opt_states, rng, env_state, timestep)
             return learner_state, transition
-
-        # INITIALISE RNN STATE
-        initial_hstates = learner_state.hstates
 
         # STEP ENVIRONMENT FOR ROLLOUT LENGTH
         learner_state, traj_batch = jax.lax.scan(
@@ -152,30 +114,8 @@ def get_learner_fn(
         )
 
         # CALCULATE ADVANTAGE
-        (
-            params,
-            opt_states,
-            rng,
-            env_state,
-            last_timestep,
-            last_done,
-            hstates,
-        ) = learner_state
-
-        # Add a batch dimension to the observation.
-        batched_last_observation = jax.tree_util.tree_map(
-            lambda x: x[jnp.newaxis, :], last_timestep.observation
-        )
-        ac_in = (
-            batched_last_observation,
-            last_done[:, 0][jnp.newaxis, :],
-        )
-
-        # Run the network.
-        _, last_val = critic_apply_fn(params.critic_params, hstates.critic_hidden_state, ac_in)
-        # Squeeze out the batch dimension and mask out the value of terminal states.
-        last_val = last_val.squeeze(0)
-        last_val = jnp.where(last_done, jnp.zeros_like(last_val), last_val)
+        params, opt_states, rng, env_state, last_timestep = learner_state
+        last_val = critic_apply_fn(params.critic_params, last_timestep.observation)
 
         def _calculate_gae(
             traj_batch: PPOTransition, last_val: chex.Array
@@ -212,14 +152,9 @@ def get_learner_fn(
             def _update_minibatch(train_state: Tuple, batch_info: Tuple) -> Tuple:
                 """Update the network for a single minibatch."""
 
+                # UNPACK TRAIN STATE AND BATCH INFO
                 params, opt_states = train_state
-                (
-                    init_policy_hstate,
-                    init_critic_hstate,
-                    traj_batch,
-                    advantages,
-                    targets,
-                ) = batch_info
+                traj_batch, advantages, targets = batch_info
 
                 def _actor_loss_fn(
                     actor_params: FrozenDict,
@@ -229,13 +164,10 @@ def get_learner_fn(
                 ) -> Tuple:
                     """Calculate the actor loss."""
                     # RERUN NETWORK
-
-                    obs_and_done = (traj_batch.obs, traj_batch.done[:, :, 0])
-                    _, actor_policy = actor_apply_fn(
-                        actor_params, init_policy_hstate.squeeze(0), obs_and_done
-                    )
+                    actor_policy = actor_apply_fn(actor_params, traj_batch.obs)
                     log_prob = actor_policy.log_prob(traj_batch.action)
 
+                    # CALCULATE ACTOR LOSS
                     ratio = jnp.exp(log_prob - traj_batch.log_prob)
                     gae = (gae - gae.mean()) / (gae.std() + 1e-8)
                     loss_actor1 = ratio * gae
@@ -251,8 +183,8 @@ def get_learner_fn(
                     loss_actor = loss_actor.mean()
                     entropy = actor_policy.entropy().mean()
 
-                    total_loss = loss_actor - config.system.ent_coef * entropy
-                    return total_loss, (loss_actor, entropy)
+                    total_loss_actor = loss_actor - config.system.ent_coef * entropy
+                    return total_loss_actor, (loss_actor, entropy)
 
                 def _critic_loss_fn(
                     critic_params: FrozenDict,
@@ -262,10 +194,7 @@ def get_learner_fn(
                 ) -> Tuple:
                     """Calculate the critic loss."""
                     # RERUN NETWORK
-                    obs_and_done = (traj_batch.obs, traj_batch.done[:, :, 0])
-                    _, value = critic_apply_fn(
-                        critic_params, init_critic_hstate.squeeze(0), obs_and_done
-                    )
+                    value = critic_apply_fn(critic_params, traj_batch.obs)
 
                     # CALCULATE VALUE LOSS
                     value_pred_clipped = traj_batch.value + (value - traj_batch.value).clip(
@@ -275,8 +204,8 @@ def get_learner_fn(
                     value_losses_clipped = jnp.square(value_pred_clipped - targets)
                     value_loss = 0.5 * jnp.maximum(value_losses, value_losses_clipped).mean()
 
-                    total_loss = config.system.vf_coef * value_loss
-                    return total_loss, (value_loss)
+                    critic_total_loss = config.system.vf_coef * value_loss
+                    return critic_total_loss, (value_loss)
 
                 # CALCULATE ACTOR LOSS
                 actor_grad_fn = jax.value_and_grad(_actor_loss_fn, has_aux=True)
@@ -337,78 +266,43 @@ def get_learner_fn(
 
                 return (new_params, new_opt_state), loss_info
 
-            (
-                params,
-                opt_states,
-                init_hstates,
-                traj_batch,
-                advantages,
-                targets,
-                rng,
-            ) = update_state
-            init_policy_hstate, init_critic_hstate = init_hstates
+            params, opt_states, traj_batch, advantages, targets, rng = update_state
             rng, shuffle_rng = jax.random.split(rng)
 
             # SHUFFLE MINIBATCHES
-            permutation = jax.random.permutation(shuffle_rng, config.arch.num_envs)
-            batch = (init_policy_hstate, init_critic_hstate, traj_batch, advantages, targets)
+            batch_size = config.system.rollout_length * config.arch.num_envs
+            permutation = jax.random.permutation(shuffle_rng, batch_size)
+            batch = (traj_batch, advantages, targets)
+            batch = jax.tree_util.tree_map(lambda x: merge_leading_dims(x, 2), batch)
             shuffled_batch = jax.tree_util.tree_map(
-                lambda x: jnp.take(x, permutation, axis=1), batch
+                lambda x: jnp.take(x, permutation, axis=0), batch
             )
-            reshaped_batch = jax.tree_util.tree_map(
-                lambda x: jnp.reshape(
-                    x, (x.shape[0], config.system.num_minibatches, -1, *x.shape[2:])
-                ),
+            minibatches = jax.tree_util.tree_map(
+                lambda x: jnp.reshape(x, [config.system.num_minibatches, -1] + list(x.shape[1:])),
                 shuffled_batch,
             )
-            minibatches = jax.tree_util.tree_map(lambda x: jnp.swapaxes(x, 1, 0), reshaped_batch)
 
             # UPDATE MINIBATCHES
             (params, opt_states), loss_info = jax.lax.scan(
                 _update_minibatch, (params, opt_states), minibatches
             )
 
-            update_state = (
-                params,
-                opt_states,
-                init_hstates,
-                traj_batch,
-                advantages,
-                targets,
-                rng,
-            )
+            update_state = (params, opt_states, traj_batch, advantages, targets, rng)
             return update_state, loss_info
 
-        init_hstates = jax.tree_util.tree_map(lambda x: x[None, :], initial_hstates)
-        update_state = (
-            params,
-            opt_states,
-            init_hstates,
-            traj_batch,
-            advantages,
-            targets,
-            rng,
-        )
+        update_state = (params, opt_states, traj_batch, advantages, targets, rng)
 
         # UPDATE EPOCHS
         update_state, loss_info = jax.lax.scan(
             _update_epoch, update_state, None, config.system.ppo_epochs
         )
 
-        params, opt_states, _, traj_batch, advantages, targets, rng = update_state
-        learner_state = RNNLearnerState(
-            params,
-            opt_states,
-            rng,
-            env_state,
-            last_timestep,
-            last_done,
-            hstates,
-        )
+        params, opt_states, traj_batch, advantages, targets, rng = update_state
+        learner_state = LearnerState(params, opt_states, rng, env_state, last_timestep)
         metric = traj_batch.info
         return learner_state, (metric, loss_info)
 
-    def learner_fn(learner_state: RNNLearnerState) -> ExperimentOutput[RNNLearnerState]:
+    def learner_fn(learner_state: LearnerState) -> ExperimentOutput[LearnerState]:
         """Learner function.
 
         This function represents the learner, it updates the network parameters
@@ -422,8 +316,6 @@ def get_learner_fn(
                 - rng (chex.PRNGKey): The random number generator state.
                 - env_state (LogEnvState): The environment state.
                 - timesteps (TimeStep): The initial timestep in the initial trajectory.
-                - dones (bool): Whether the initial timestep was a terminal state.
-                - hstateS (HiddenStates): The initial hidden states of the RNN.
         """
 
         batched_update_step = jax.vmap(_update_step, in_axes=(0, None), axis_name="batch")
@@ -446,7 +338,7 @@ def get_learner_fn(
 
 def learner_setup(
     env: Environment, rngs: chex.Array, config: DictConfig
-) -> Tuple[LearnerFn[RNNLearnerState], Actor, RNNLearnerState]:
+) -> Tuple[LearnerFn[LearnerState], Actor, LearnerState]:
     """Initialise learner_fn, network, optimiser, environment and states."""
     # Get available TPU cores.
     n_devices = len(jax.devices())
@@ -460,9 +352,11 @@ def learner_setup(
     # PRNG keys.
     rng, rng_p = rngs
 
-    # Define network and optimisers.
+    # Define network and optimiser.
     actor_network, critic_network = get_networks(
-        config=config, network="recurrent", centralised_critic=False
+        config=config,
+        network="feedforward",
+        centralised_critic=True,
     )
     actor_optim = optax.chain(
         optax.clip_by_global_norm(config.system.max_grad_norm),
@@ -473,42 +367,38 @@ def learner_setup(
         optax.adam(config.system.critic_lr, eps=1e-5),
     )
 
-    # Initialise observation: Select only obs for a single agent.
-    init_obs = env.observation_spec().generate_value()
-    init_obs = jax.tree_util.tree_map(lambda x: x[0], init_obs)
-    init_obs = jax.tree_util.tree_map(
-        lambda x: jnp.repeat(x[jnp.newaxis, ...], config.arch.num_envs, axis=0),
-        init_obs,
+    # Initialise observation.
+    obs = env.observation_spec().generate_value()
+    # Select only obs for a single agent.
+    init_x = ObservationGlobalState(
+        agents_view=obs.agents_view[0],
+        action_mask=obs.action_mask[0],
+        global_state=obs.global_state[0],
+        step_count=obs.step_count[0],
     )
-    init_obs = jax.tree_util.tree_map(lambda x: x[None, ...], init_obs)
-    init_done = jnp.zeros((1, config.arch.num_envs), dtype=bool)
-    init_x = (init_obs, init_done)
+    init_x = jax.tree_util.tree_map(lambda x: x[None, ...], init_x)
 
-    # Initialise hidden states.
-    init_policy_hstate = ScannedRNN.initialize_carry(
-        (config.arch.num_envs), config.system.actor_network.pre_torso_layer_sizes[-1]
-    )
-    init_critic_hstate = ScannedRNN.initialize_carry(
-        (config.arch.num_envs),
-        config.system.critic_network.pre_torso_layer_sizes[-1],
-    )
-
-    # initialise params and optimiser state.
-    actor_params = actor_network.init(rng_p, init_policy_hstate, init_x)
+    # Initialise actor params and optimiser state.
+    actor_params = actor_network.init(rng_p, init_x)
     actor_opt_state = actor_optim.init(actor_params)
-    critic_params = critic_network.init(rng_p, init_critic_hstate, init_x)
+
+    # Initialise critic params and optimiser state.
+    critic_params = critic_network.init(rng_p, init_x)
     critic_opt_state = critic_optim.init(critic_params)
 
     # Vmap network apply function over number of agents.
     vmapped_actor_network_apply_fn = jax.vmap(
-        actor_network.apply, in_axes=(None, 1, (2, None)), out_axes=(1, 2)
+        actor_network.apply,
+        in_axes=(None, 1),
+        out_axes=(1),
     )
-    # Vmap network apply function over number of agents.
     vmapped_critic_network_apply_fn = jax.vmap(
-        critic_network.apply, in_axes=(None, 1, (2, None)), out_axes=(1, 2)
+        critic_network.apply,
+        in_axes=(None, 1),
+        out_axes=(1),
     )
 
-    # Get network apply functions and optimiser updates.
+    # Pack apply and update functions.
     apply_fns = (vmapped_actor_network_apply_fn, vmapped_critic_network_apply_fn)
     update_fns = (actor_optim.update, critic_optim.update)
 
@@ -525,15 +415,6 @@ def learner_setup(
     critic_params = jax.tree_map(broadcast, critic_params)
     critic_opt_state = jax.tree_map(broadcast, critic_opt_state)
 
-    # Duplicate the hidden state for each agent.
-    init_policy_hstate = jnp.expand_dims(init_policy_hstate, axis=1)
-    init_policy_hstate = jnp.tile(init_policy_hstate, (1, config.system.num_agents, 1))
-    policy_hstates = jax.tree_map(broadcast, init_policy_hstate)
-
-    init_critic_hstate = jnp.expand_dims(init_critic_hstate, axis=1)
-    init_critic_hstate = jnp.tile(init_critic_hstate, (1, config.system.num_agents, 1))
-    critic_hstates = jax.tree_map(broadcast, init_critic_hstate)
-
     # Initialise environment states and timesteps.
     rng, *env_rngs = jax.random.split(
         rng, n_devices * config.system.update_batch_size * config.arch.num_envs + 1
@@ -544,6 +425,7 @@ def learner_setup(
 
     # Split rngs for each core.
     rng, *step_rngs = jax.random.split(rng, n_devices * config.system.update_batch_size + 1)
+
     # Add dimension to pmap over.
     reshape_step_rngs = lambda x: x.reshape(
         (n_devices, config.system.update_batch_size) + x.shape[1:]
@@ -555,28 +437,10 @@ def learner_setup(
     env_states = jax.tree_util.tree_map(reshape_states, env_states)
     timesteps = jax.tree_util.tree_map(reshape_states, timesteps)
 
-    # Initialise dones.
-    dones = jnp.zeros(
-        (
-            n_devices,
-            config.system.update_batch_size,
-            config.arch.num_envs,
-            config.system.num_agents,
-        ),
-        dtype=bool,
-    )
-    hstates = HiddenStates(policy_hstates, critic_hstates)
     params = Params(actor_params, critic_params)
     opt_states = OptStates(actor_opt_state, critic_opt_state)
-    init_learner_state = RNNLearnerState(
-        params=params,
-        opt_states=opt_states,
-        key=step_rngs,
-        env_state=env_states,
-        timestep=timesteps,
-        dones=dones,
-        hstates=hstates,
-    )
+
+    init_learner_state = LearnerState(params, opt_states, step_rngs, env_states, timesteps)
     return learn, actor_network, init_learner_state
 
 
@@ -604,13 +468,10 @@ def run_experiment(_config: DictConfig) -> None:
         network=actor_network,
         params=learner_state.params.actor_params,
         config=config,
-        use_recurrent_net=True,
-        scanned_rnn=ScannedRNN,
     )
 
     # Calculate total timesteps.
     n_devices = len(jax.devices())
-
     config.system.num_updates_per_eval = config.system.num_updates // config.arch.num_evaluation
     steps_per_rollout = (
         n_devices
@@ -671,7 +532,8 @@ def run_experiment(_config: DictConfig) -> None:
         # Prepare for evaluation.
         start_time = time.time()
         trained_params = jax.tree_util.tree_map(
-            lambda x: x[:, 0, ...], learner_output.learner_state.params.actor_params
+            lambda x: x[:, 0, ...],
+            learner_output.learner_state.params.actor_params,  # Select only actor params
         )
         rng_e, *eval_rngs = jax.random.split(rng_e, n_devices + 1)
         eval_rngs = jnp.stack(eval_rngs)
@@ -728,14 +590,14 @@ def run_experiment(_config: DictConfig) -> None:
     logger.stop()
 
 
-@hydra.main(config_path="../configs", config_name="default_rec_ippo.yaml", version_base="1.2")
+@hydra.main(config_path="../../configs", config_name="default_ff_mappo.yaml", version_base="1.2")
 def hydra_entry_point(cfg: DictConfig) -> None:
     """Experiment entry point."""
 
     # Run experiment.
     run_experiment(cfg)
 
-    print(f"{Fore.CYAN}{Style.BRIGHT}Recurrent IPPO experiment completed{Style.RESET_ALL}")
+    print(f"{Fore.CYAN}{Style.BRIGHT}MAPPO experiment completed{Style.RESET_ALL}")
 
 
 if __name__ == "__main__":
